feat: add AppContent
